+++ software/nn/src/core/tensor/tests/gpu_buffer_pool_gtest.cpp
@@ -8,7 +8,7 @@
 
 #include <vector>
 
-#include "nn/tensor/GPUBufferPool.hpp"
+#include "nn/tensor/opencl/GPUBufferPool.hpp"
 
 using namespace nn::tensor;